refactor: split MCP tools into individual modules
Split the 852-line mcp-server.ts monolith into separate tool files:
- mcp/channel.ts (channel_send, channel_read)
- mcp/resource.ts (resource_create, resource_read)
- mcp/inbox.ts (my_inbox, my_inbox_ack, my_status_set)
- mcp/team.ts (team_members, team_doc_*)
- mcp/proposal.ts (team_proposal_*, team_vote)
- mcp/feedback.ts (feedback_submit)
- mcp/skills.ts (skills_list, skills_view, skills_read) — new MCP version
- mcp/server.ts — thin orchestrator
- mcp/helpers.ts, mcp/types.ts — shared deps

Tools are now clearly split:
  SDK tools (agent/tools/) — for ToolLoopAgent
  MCP tools (workflow/context/mcp/) — for McpServer

Skills and feedback exist in both forms (same data layer, different transport).
